Refactor integration tests of Citadel
- Citadel is a security component -> organize Citadel integration tests
under the security integration tests folder.
- The common utility functions are refactored into the util folder.
